Changed mood_analyzer, refactored comment_labeler and post_summarizer, changed gazeta and podrobno scrapers
In mood_analyzer.py, I added more keywords that seem popular because it trimmed all the of the instagram posts down to 2 out of the 50. I added the summarized posts to the prompt to the LLM.
I changed the names for post_summarizer and comment_labeler in anticipation for an article version of those files.
I added a database reset statement to the gazeta and podrobno scrapers.